@@ -14,19 +14,19 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  firebase_auth:
-  cloud_firestore:
-  firebase_storage:
-  image_picker:
+  firebase_auth: ^1.0.0  
+  cloud_firestore: ^1.0.0
+  firebase_storage: ^8.0.0
+  image_picker: ^0.7.2
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  # cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
