@@ -1,18 +1,18 @@
 name: dash_chat
 description: The most complete Chat UI for flutter inspired by react-native-gifted-chat.
-version: 1.1.15
+version: 1.1.15-nullsafety
 homepage: https://github.com/fayeed/dash_chat
 
 environment:
-  sdk: '>=2.2.2 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_parsed_text: ^1.1.0
-  uuid: ^3.0.4
+  flutter_parsed_text: ^2.1.0
+  uuid: ^3.0.1
   intl: ^0.17.0
-  transparent_image: ^1.0.0
+  transparent_image: ^2.0.0-nullsafety.0
 
 dev_dependencies:
   flutter_test:
