@@ -13,10 +13,10 @@ class DashChat extends StatefulWidget {
   final bool readOnly;
 
   /// Height for the Dash chat Widget
-  final double height;
+  final double? height;
 
   // Width for the Dash chat Widget
-  final double width;
+  final double? width;
 
   /// List of messages to display in the chat container
   /// Takes a [List] of [ChatMessage]
@@ -24,10 +24,10 @@ class DashChat extends StatefulWidget {
 
   /// If provided, this text editing controller will be used for
   /// the text input.
-  final TextEditingController textController;
+  final TextEditingController? textController;
 
   /// If provided, this focus node will be used for the text input.
-  final FocusNode focusNode;
+  final FocusNode? focusNode;
 
   /// Use to change the direction of the text ltr is used for
   /// launguages that start from left like English &
@@ -39,11 +39,11 @@ class DashChat extends StatefulWidget {
   /// If provided will stop using the default controller
   /// i.e [TextEditingController] and will use this to update the
   /// text input field.
-  final String text;
+  final String? text;
 
   /// If the text parameter is passed then onTextChange must also
   /// be passed.
-  final Function(String) onTextChange;
+  final Function(String)? onTextChange;
 
   /// If the input TextField is disabled.
   final bool inputDisabled;
@@ -51,7 +51,7 @@ class DashChat extends StatefulWidget {
   /// Used to provide input decoration to the text as default only
   /// to the input placeholder for the chat input
   /// "Add Message here...".
-  final InputDecoration inputDecoration;
+  final InputDecoration? inputDecoration;
 
   ///Configures how the platform keyboard will select an uppercase or lowercase keyboard.
   ///Only supports text keyboards, other keyboard types will ignore this configuration. Capitalization is locale-aware.
@@ -60,7 +60,7 @@ class DashChat extends StatefulWidget {
 
   /// Usually new message added by the user gets [Uuid] String
   /// Can be override by proving this parameter
-  final String Function() messageIdGenerator;
+  final String Function()? messageIdGenerator;
 
   /// The current user object [ChatUser].
   final ChatUser user;
@@ -78,15 +78,15 @@ class DashChat extends StatefulWidget {
   final bool sendOnEnter;
 
   /// Input action of the keyboard
-  final TextInputAction textInputAction;
+  final TextInputAction? textInputAction;
 
   /// [DateFormat] object for formatting date to show in [MessageListView]
   /// defaults to `yyyy-MM-dd`.
-  final DateFormat dateFormat;
+  final DateFormat? dateFormat;
 
   /// [DateFormat] object for formatting time to show in [MessageContainer]
   /// defaults to `HH:mm:ss`.
-  final DateFormat timeFormat;
+  final DateFormat? timeFormat;
 
   /// Should the user avatar be shown defaults to false and will not
   /// show the user avatar.
@@ -94,7 +94,7 @@ class DashChat extends StatefulWidget {
 
   /// avatarBuilder will override the the default avatar which uses
   /// [CircleAvatar].
-  final Widget Function(ChatUser) avatarBuilder;
+  final Widget Function(ChatUser)? avatarBuilder;
 
   /// Should the avatar be shown for every message defaulst to false.
   final bool showAvatarForEveryMessage;
@@ -102,17 +102,17 @@ class DashChat extends StatefulWidget {
   /// [onPressAvatar] function takes a function with this structure
   /// [Function(ChatUser)] will trigger when the avatar
   /// is tapped on
-  final Function(ChatUser) onPressAvatar;
+  final Function(ChatUser)? onPressAvatar;
 
   /// [onLongPressAvatar] function takea a function with this structure
   /// [Function(ChatUser)] will trigger when the avatar
   /// is long pressed
-  final Function(ChatUser) onLongPressAvatar;
+  final Function(ChatUser)? onLongPressAvatar;
 
   /// [onLongPressMessage] function takea a function with this structure
   /// [Function(ChatMessage)] will trigger when the message
   /// is long pressed.
-  final Function(ChatMessage) onLongPressMessage;
+  final Function(ChatMessage)? onLongPressMessage;
 
   /// Should the messages be shown in reversed order.
   final bool inverted;
@@ -120,26 +120,27 @@ class DashChat extends StatefulWidget {
   /// messageBuilder will override the the default chat container which uses
   /// and you will need to build complete message Widget it will not accept
   /// and include any other builder functions.
-  final Widget Function(ChatMessage) messageBuilder;
+  final Widget Function(ChatMessage)? messageBuilder;
 
   /// messageTextBuilder will override the the default message text.
-  final Widget Function(String, [ChatMessage]) messageTextBuilder;
+  final Widget Function(String?, [ChatMessage])? messageTextBuilder;
 
   /// messageImageBuilder will override the the default Image.
-  final Widget Function(String url, [ChatMessage]) messageImageBuilder;
+  final Widget Function(String? url, [ChatMessage])? messageImageBuilder;
 
   /// messageTimeBuilder will override the the default text.
-  final Widget Function(String url, [ChatMessage]) messageTimeBuilder;
+  final Widget Function(String formattedTime, [ChatMessage])?
+      messageTimeBuilder;
 
   /// dateBuilder will override the the default time text.
-  final Widget Function(String) dateBuilder;
+  final Widget Function(String)? dateBuilder;
 
   /// A Widget that will be shown below the [MessageListView] like you can
   /// show a "tying..." at the end.
-  final Widget Function() chatFooterBuilder;
+  final Widget Function()? chatFooterBuilder;
 
   /// Main input length of the input text box defaulst to no limit.
-  final int maxInputLength;
+  final int? maxInputLength;
 
   /// Used to parse text to make it linkified text uses
   /// [flutter_parsed_text](https://pub.dev/packages/flutter_parsed_text)
@@ -149,7 +150,7 @@ class DashChat extends StatefulWidget {
 
   /// Provides a custom style to the message container
   /// takes [BoxDecoration]
-  final BoxDecoration messageContainerDecoration;
+  final BoxDecoration? messageContainerDecoration;
 
   /// [List] of [Widget] to show before the [TextField].
   final List<Widget> leading;
@@ -159,13 +160,13 @@ class DashChat extends StatefulWidget {
   final List<Widget> trailing;
 
   /// sendButtonBuilder will override the the default [IconButton].
-  final Widget Function(Function) sendButtonBuilder;
+  final Widget Function(Function)? sendButtonBuilder;
 
   /// Style for the [TextField].
-  final TextStyle inputTextStyle;
+  final TextStyle? inputTextStyle;
 
   /// [TextField] container style.
-  final BoxDecoration inputContainerStyle;
+  final BoxDecoration? inputContainerStyle;
 
   /// Max length of the input lines default to 1.
   final int inputMaxLines;
@@ -177,35 +178,35 @@ class DashChat extends StatefulWidget {
   final double inputCursorWidth;
 
   /// Color of the input cursor defaults to theme.
-  final Color inputCursorColor;
+  final Color? inputCursorColor;
 
   /// ScrollController for the [MessageListView] will use the default
   /// scrollcontroller in the Widget.
-  final ScrollController scrollController;
+  final ScrollController? scrollController;
 
   /// A Widget that will be shown below the [ChatInputToolbar] like you can
   /// show a list of buttons like file image just like in Slack app.
-  final Widget Function() inputFooterBuilder;
+  final Widget Function()? inputFooterBuilder;
 
   /// Padding for the [MessageListView].
   final EdgeInsetsGeometry messageContainerPadding;
 
   /// Callback method when the quickReply was tapped on
   /// will pass [Reply] as a paramter to function.
-  final Function(Reply) onQuickReply;
+  final Function(Reply)? onQuickReply;
 
   /// Padding for the quick reply area
   /// by default it padding is set 0.0
   final EdgeInsetsGeometry quickReplyPadding;
 
   /// Container style for the QuickReply Container [BoxDecoration].
-  final BoxDecoration quickReplyStyle;
+  final BoxDecoration? quickReplyStyle;
 
   /// [TextStyle] for QuickReply textstyle.
-  final TextStyle quickReplyTextStyle;
+  final TextStyle? quickReplyTextStyle;
 
   /// quickReplyBuilder will override the the default QuickReply Widget.
-  final Widget Function(Reply) quickReplyBuilder;
+  final Widget Function(Reply)? quickReplyBuilder;
 
   /// Should quick reply be horizontally scrollable
   final bool quickReplyScroll;
@@ -221,10 +222,10 @@ class DashChat extends StatefulWidget {
   final bool shouldStartMessagesFromTop;
 
   /// Overrides the default [scrollToBottomWidget] with a custom widget
-  final Widget Function() scrollToBottomWidget;
+  final Widget Function()? scrollToBottomWidget;
 
   /// Override the default behaviour of the onScrollToBottom Widget
-  final Function onScrollToBottomPress;
+  final Function? onScrollToBottomPress;
 
   /// Should the LoadEarlier Floating widget be shown or use
   /// load as you scroll scheme whcih will call the [onLoadEarlier]
@@ -234,11 +235,11 @@ class DashChat extends StatefulWidget {
   final bool shouldShowLoadEarlier;
 
   /// Override the default behaviour of the onScrollToBottom Widget
-  final Widget Function() showLoadEarlierWidget;
+  final Widget Function()? showLoadEarlierWidget;
 
   /// Override the default behaviour of the onLoadEarleir Widget
   /// or used as a callback when the listView reaches the top
-  final Function onLoadEarlier;
+  final Function? onLoadEarlier;
 
   /// Padding for the default input toolbar
   /// by default it padding is set 0.0
@@ -251,7 +252,7 @@ class DashChat extends StatefulWidget {
   /// [messageButtonsBuilder] function takes a function with this
   /// structure [List<Widget> Function()] to render the buttons inside
   /// a row.
-  final List<Widget> Function(ChatMessage) messageButtonsBuilder;
+  final List<Widget> Function(ChatMessage)? messageButtonsBuilder;
 
   /// Padding of the message
   /// Default to EdgeInsets.all(8.0)
@@ -271,13 +272,13 @@ class DashChat extends StatefulWidget {
   /// can be used to override color, or customise the message container
   /// params [ChatMessage] and [isUser]: boolean
   /// return BoxDecoration
-  final BoxDecoration Function(ChatMessage, bool) messageDecorationBuilder;
+  final BoxDecoration Function(ChatMessage, bool?)? messageDecorationBuilder;
 
-  ScrollToBottomStyle scrollToBottomStyle;
+  late ScrollToBottomStyle scrollToBottomStyle;
 
   DashChat({
-    Key key,
-    ScrollToBottomStyle scrollToBottomStyle,
+    Key? key,
+    ScrollToBottomStyle? scrollToBottomStyle,
     this.avatarMaxSize = 30.0,
     this.inputTextDirection = TextDirection.ltr,
     this.inputToolbarMargin = const EdgeInsets.all(0.0),
@@ -314,7 +315,7 @@ class DashChat extends StatefulWidget {
     this.height,
     this.width,
     this.readOnly = false,
-    @required this.messages,
+    required this.messages,
     this.onTextChange,
     this.text,
     this.inputDisabled = false,
@@ -326,8 +327,8 @@ class DashChat extends StatefulWidget {
     this.messageIdGenerator,
     this.dateFormat,
     this.timeFormat,
-    @required this.user,
-    @required this.onSend,
+    required this.user,
+    required this.onSend,
     this.onLongPressAvatar,
     this.onLongPressMessage,
     this.onPressAvatar,
@@ -355,7 +356,7 @@ class DashChat extends StatefulWidget {
     this.scrollToBottomStyle = scrollToBottomStyle ?? new ScrollToBottomStyle();
   }
 
-  String getVal() {
+  String? getVal() {
     return text;
   }
 
@@ -364,9 +365,9 @@ class DashChat extends StatefulWidget {
 }
 
 class DashChatState extends State<DashChat> {
-  FocusNode inputFocusNode;
-  TextEditingController textController;
-  ScrollController scrollController;
+  FocusNode? inputFocusNode;
+  TextEditingController? textController;
+  late ScrollController scrollController;
   String _text = "";
   bool visible = false;
   GlobalKey inputKey = GlobalKey();
@@ -374,7 +375,7 @@ class DashChatState extends State<DashChat> {
   bool showLoadMore = false;
   String get messageInput => _text;
   bool _initialLoad = true;
-  Timer _timer;
+  Timer? _timer;
 
   void onTextChange(String text) {
     if (visible) {
@@ -404,7 +405,7 @@ class DashChatState extends State<DashChat> {
     scrollController = widget.scrollController ?? ScrollController();
     textController = widget.textController ?? TextEditingController();
     inputFocusNode = widget.focusNode ?? FocusNode();
-    WidgetsBinding.instance.addPostFrameCallback(widgetBuilt);
+    WidgetsBinding.instance!.addPostFrameCallback(widgetBuilt);
     super.initState();
   }
 
@@ -455,7 +456,7 @@ class DashChatState extends State<DashChat> {
           });
         }
       } else if (topReached) {
-        widget.onLoadEarlier();
+        widget.onLoadEarlier!();
       }
     });
   }
@@ -481,42 +482,44 @@ class DashChatState extends State<DashChat> {
                     : MainAxisAlignment.end,
                 children: <Widget>[
                   MessageListView(
-                    avatarMaxSize: widget.avatarMaxSize,
-                    messagePadding: widget.messagePadding,
-                    constraints: constraints,
-                    shouldShowLoadEarlier: widget.shouldShowLoadEarlier,
-                    showLoadEarlierWidget: widget.showLoadEarlierWidget,
-                    onLoadEarlier: widget.onLoadEarlier,
-                    defaultLoadCallback: changeDefaultLoadMore,
-                    messageContainerPadding: widget.messageContainerPadding,
-                    scrollController: widget.scrollController != null
-                        ? widget.scrollController
-                        : scrollController,
-                    user: widget.user,
-                    messages: widget.messages,
-                    showuserAvatar: widget.showUserAvatar,
-                    dateFormat: widget.dateFormat,
-                    timeFormat: widget.timeFormat,
-                    inverted: widget.inverted,
-                    showAvatarForEverMessage: widget.showAvatarForEveryMessage,
-                    onLongPressAvatar: widget.onLongPressAvatar,
-                    onPressAvatar: widget.onPressAvatar,
-                    onLongPressMessage: widget.onLongPressMessage,
-                    avatarBuilder: widget.avatarBuilder,
-                    messageBuilder: widget.messageBuilder,
-                    messageTextBuilder: widget.messageTextBuilder,
-                    messageImageBuilder: widget.messageImageBuilder,
-                    messageTimeBuilder: widget.messageTimeBuilder,
-                    dateBuilder: widget.dateBuilder,
-                    messageContainerDecoration:
-                        widget.messageContainerDecoration,
-                    parsePatterns: widget.parsePatterns,
-                    changeVisible: changeVisible,
-                    visible: visible,
-                    showLoadMore: showLoadMore,
-                    messageButtonsBuilder: widget.messageButtonsBuilder,
-                    messageDecorationBuilder: widget.messageDecorationBuilder
-                  ),
+                      avatarMaxSize: widget.avatarMaxSize,
+                      messagePadding: widget.messagePadding,
+                      constraints: constraints,
+                      shouldShowLoadEarlier: widget.shouldShowLoadEarlier,
+                      showLoadEarlierWidget: widget.showLoadEarlierWidget,
+                      onLoadEarlier: widget.onLoadEarlier,
+                      defaultLoadCallback: changeDefaultLoadMore,
+                      messageContainerPadding:
+                          widget.messageContainerPadding as EdgeInsets,
+                      scrollController: widget.scrollController != null
+                          ? widget.scrollController
+                          : scrollController,
+                      user: widget.user,
+                      messages: widget.messages,
+                      showuserAvatar: widget.showUserAvatar,
+                      dateFormat: widget.dateFormat,
+                      timeFormat: widget.timeFormat,
+                      inverted: widget.inverted,
+                      showAvatarForEverMessage:
+                          widget.showAvatarForEveryMessage,
+                      onLongPressAvatar: widget.onLongPressAvatar,
+                      onPressAvatar: widget.onPressAvatar,
+                      onLongPressMessage: widget.onLongPressMessage,
+                      avatarBuilder: widget.avatarBuilder,
+                      messageBuilder: widget.messageBuilder,
+                      messageTextBuilder: widget.messageTextBuilder,
+                      messageImageBuilder: widget.messageImageBuilder,
+                      messageTimeBuilder: widget.messageTimeBuilder,
+                      dateBuilder: widget.dateBuilder,
+                      messageContainerDecoration:
+                          widget.messageContainerDecoration,
+                      parsePatterns: widget.parsePatterns,
+                      changeVisible: changeVisible,
+                      visible: visible,
+                      showLoadMore: showLoadMore,
+                      messageButtonsBuilder: widget.messageButtonsBuilder,
+                      messageDecorationBuilder:
+                          widget.messageDecorationBuilder),
                   if (widget.messages.length != 0 &&
                       widget.messages.last.user.uid != widget.user.uid &&
                       widget.messages.last.quickReplies != null)
@@ -528,20 +531,21 @@ class DashChatState extends State<DashChat> {
                       child: widget.quickReplyScroll
                           ? ListView(
                               scrollDirection: Axis.horizontal,
-                              children: widget.messages.last.quickReplies.values
+                              children: widget
+                                  .messages.last.quickReplies!.values!
                                   .map(_mapReply)
                                   .toList(),
                             )
                           : Wrap(
                               children: <Widget>[
-                                ...widget.messages.last.quickReplies.values
+                                ...widget.messages.last.quickReplies!.values!
                                     .sublist(
                                         0,
-                                        widget.messages.last.quickReplies.values
-                                                    .length <=
+                                        widget.messages.last.quickReplies!
+                                                    .values!.length <=
                                                 3
-                                            ? widget.messages.last.quickReplies
-                                                .values.length
+                                            ? widget.messages.last.quickReplies!
+                                                .values!.length
                                             : 3)
                                     .map(_mapReply)
                                     .toList(),
@@ -549,7 +553,7 @@ class DashChatState extends State<DashChat> {
                             ),
                     ),
                   if (widget.chatFooterBuilder != null)
-                    widget.chatFooterBuilder(),
+                    widget.chatFooterBuilder!(),
                   if (!widget.readOnly)
                     SafeArea(
                       child: ChatInputToolbar(
@@ -599,7 +603,7 @@ class DashChatState extends State<DashChat> {
                   bottom: widget.scrollToBottomStyle.bottom,
                   top: widget.scrollToBottomStyle.top,
                   child: widget.scrollToBottomWidget != null
-                      ? widget.scrollToBottomWidget()
+                      ? widget.scrollToBottomWidget!()
                       : ScrollToBottom(
                           onScrollToBottomPress: widget.onScrollToBottomPress,
                           scrollToBottomStyle: widget.scrollToBottomStyle,
