@@ -4,10 +4,10 @@ part of dash_chat;
 /// be shown to the user
 class QuickReplies {
   /// [List] of replies that will be shown to the user
-  List<Reply> values;
+  List<Reply>? values;
 
   /// Should the quick replies be dismissable or persist
-  bool keepIt;
+  bool? keepIt;
 
   QuickReplies({
     this.keepIt,
@@ -15,10 +15,10 @@ class QuickReplies {
   });
 
   QuickReplies.fromJson(Map<dynamic, dynamic> json) {
-    keepIt = json['keepIt'] != null ? json['keepIt'] : null;
+    keepIt = json['keepIt'];
 
     if (json['values'] != null) {
-      List<Reply> replies = List<Reply>();
+      List<Reply> replies = <Reply>[];
 
       for (var i = 0; i < json['values'].length; i++) {
         replies.add(Reply.fromJson(json['values'][i]));
@@ -32,7 +32,7 @@ class QuickReplies {
     final Map<String, dynamic> data = Map<String, dynamic>();
 
     data['keepIt'] = keepIt;
-    data['values'] = values.map((e) => e.toJson()).toList();
+    data['values'] = values!.map((e) => e.toJson()).toList();
 
     return data;
   }
