@@ -11,25 +11,25 @@ class AvatarContainer extends StatelessWidget {
   /// [onPress] function takea a function with this structure
   /// [Function(ChatUser)] will trigger when the avatar
   /// is tapped on
-  final Function(ChatUser) onPress;
+  final Function(ChatUser)? onPress;
 
   /// [onLongPress] function takea a function with this structure
   /// [Function(ChatUser)] will trigger when the avatar
   /// is long pressed
-  final Function(ChatUser) onLongPress;
+  final Function(ChatUser)? onLongPress;
 
   /// [avatarBuilder] function takea a function with this structure
   /// [Widget Function(ChatUser)] to build the avatar
-  final Widget Function(ChatUser) avatarBuilder;
+  final Widget Function(ChatUser)? avatarBuilder;
 
   /// [constraints] to apply to build the layout
   /// by default used MediaQuery and take screen size as constaints
-  final BoxConstraints constraints;
+  final BoxConstraints? constraints;
 
-  final double avatarMaxSize;
+  final double? avatarMaxSize;
 
   const AvatarContainer({
-    @required this.user,
+    required this.user,
     this.onPress,
     this.onLongPress,
     this.avatarBuilder,
@@ -45,10 +45,10 @@ class AvatarContainer extends StatelessWidget {
             maxWidth: MediaQuery.of(context).size.width);
 
     return GestureDetector(
-      onTap: () => onPress != null ? onPress(user) : null,
-      onLongPress: () => onLongPress != null ? onLongPress(user) : null,
+      onTap: () => onPress?.call(user),
+      onLongPress: () => onLongPress?.call(user),
       child: avatarBuilder != null
-          ? avatarBuilder(user)
+          ? avatarBuilder!(user)
           : Stack(
               alignment: Alignment.center,
               children: <Widget>[
@@ -57,21 +57,21 @@ class AvatarContainer extends StatelessWidget {
                     height: constraints.maxWidth * 0.08,
                     width: constraints.maxWidth * 0.08,
                     constraints: BoxConstraints(
-                      maxWidth: avatarMaxSize,
-                      maxHeight: avatarMaxSize,
+                      maxWidth: avatarMaxSize!,
+                      maxHeight: avatarMaxSize!,
                     ),
                     color: Colors.grey,
                     child: Center(
-                        child: Text(user.name == null || user.name.isEmpty
+                        child: Text(user.name == null || user.name!.isEmpty
                             ? ''
-                            : user.name[0])),
+                            : user.name![0])),
                   ),
                 ),
-                user.avatar != null && user.avatar.length != 0
+                user.avatar != null && user.avatar!.length != 0
                     ? Center(
                         child: ClipOval(
                           child: FadeInImage.memoryNetwork(
-                            image: user.avatar,
+                            image: user.avatar!,
                             placeholder: kTransparentImage,
                             fit: BoxFit.cover,
                             height: constraints.maxWidth * 0.08,
