@@ -2,24 +2,20 @@ part of dash_chat;
 
 class LoadEarlierWidget extends StatelessWidget {
   const LoadEarlierWidget({
-    Key key,
-    @required this.onLoadEarlier,
-    @required this.defaultLoadCallback,
+    Key? key,
+    this.onLoadEarlier,
+    required this.defaultLoadCallback,
   }) : super(key: key);
 
-  final Function onLoadEarlier;
+  final Function? onLoadEarlier;
   final Function(bool) defaultLoadCallback;
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
       onTap: () {
-        if (onLoadEarlier != null) {
-          onLoadEarlier();
-          defaultLoadCallback(false);
-        } else {
-          defaultLoadCallback(false);
-        }
+        onLoadEarlier?.call();
+        defaultLoadCallback(false);
       },
       child: Container(
         padding: EdgeInsets.symmetric(
