@@ -10,26 +10,26 @@ class MessageContainer extends StatelessWidget {
   /// [DateFormat] object to render the date in desired
   /// format, if no format is provided it use
   /// the default `HH:mm:ss`
-  final DateFormat timeFormat;
+  final DateFormat? timeFormat;
 
   /// [messageTextBuilder] function takes a function with this
   /// structure [Widget Function(String)] to render the text inside
   /// the container.
-  final Widget Function(String, [ChatMessage]) messageTextBuilder;
+  final Widget Function(String?, [ChatMessage])? messageTextBuilder;
 
   /// [messageImageBuilder] function takes a function with this
   /// structure [Widget Function(String)] to render the image inside
   /// the container.
-  final Widget Function(String, [ChatMessage]) messageImageBuilder;
+  final Widget Function(String?, [ChatMessage])? messageImageBuilder;
 
   /// [messageTimeBuilder] function takes a function with this
   /// structure [Widget Function(String)] to render the time text inside
   /// the container.
-  final Widget Function(String, [ChatMessage]) messageTimeBuilder;
+  final Widget Function(String, [ChatMessage])? messageTimeBuilder;
 
   /// Provides a custom style to the message container
   /// takes [BoxDecoration]
-  final BoxDecoration messageContainerDecoration;
+  final BoxDecoration? messageContainerDecoration;
 
   /// Used to parse text to make it linkified text uses
   /// [flutter_parsed_text](https://pub.dev/packages/flutter_parsed_text)
@@ -42,15 +42,15 @@ class MessageContainer extends StatelessWidget {
 
   /// Provides a list of buttons to allow the usage of adding buttons to
   /// the bottom of the message
-  final List<Widget> buttons;
+  final List<Widget>? buttons;
 
   /// [messageButtonsBuilder] function takes a function with this
   /// structure [List<Widget> Function()] to render the buttons inside
   /// a row.
-  final List<Widget> Function(ChatMessage) messageButtonsBuilder;
+  final List<Widget> Function(ChatMessage)? messageButtonsBuilder;
 
   /// Constraint to use to build the message layout
-  final BoxConstraints constraints;
+  final BoxConstraints? constraints;
 
   /// Padding of the message
   /// Default to EdgeInsets.all(8.0)
@@ -65,11 +65,11 @@ class MessageContainer extends StatelessWidget {
   /// can be used to override color, or customise the message container
   /// params [ChatMessage] and [isUser]: boolean
   /// return BoxDecoration
-  final BoxDecoration Function(ChatMessage, bool) messageDecorationBuilder;
+  final BoxDecoration Function(ChatMessage, bool?)? messageDecorationBuilder;
 
   const MessageContainer({
-    @required this.message,
-    @required this.timeFormat,
+    required this.message,
+    required this.timeFormat,
     this.constraints,
     this.messageImageBuilder,
     this.messageTextBuilder,
@@ -77,7 +77,7 @@ class MessageContainer extends StatelessWidget {
     this.messageContainerDecoration,
     this.parsePatterns = const <MatchText>[],
     this.textBeforeImage = true,
-    this.isUser,
+    required this.isUser,
     this.messageButtonsBuilder,
     this.buttons,
     this.messagePadding = const EdgeInsets.all(8.0),
@@ -95,22 +95,19 @@ class MessageContainer extends StatelessWidget {
         maxWidth: constraints.maxWidth * 0.8,
       ),
       child: Container(
-        decoration: messageDecorationBuilder != null
-            ? messageDecorationBuilder(message, isUser)
-            : messageContainerDecoration != null
-                ? messageContainerDecoration.copyWith(
-                    color: message.user.containerColor != null
-                        ? message.user.containerColor
-                        : messageContainerDecoration.color,
-                  )
-                : BoxDecoration(
-                    color: message.user.containerColor != null
-                        ? message.user.containerColor
-                        : isUser
-                            ? Theme.of(context).accentColor
-                            : Color.fromRGBO(225, 225, 225, 1),
-                    borderRadius: BorderRadius.circular(5.0),
-                  ),
+        decoration: messageDecorationBuilder?.call(message, isUser) ??
+            messageContainerDecoration?.copyWith(
+              color: message.user.containerColor != null
+                  ? message.user.containerColor
+                  : messageContainerDecoration!.color,
+            ) ??
+            BoxDecoration(
+              color: message.user.containerColor ??
+                  (isUser
+                      ? Theme.of(context).accentColor
+                      : Color.fromRGBO(225, 225, 225, 1)),
+              borderRadius: BorderRadius.circular(5.0),
+            ),
         margin: EdgeInsets.only(
           bottom: 5.0,
         ),
@@ -134,40 +131,34 @@ class MessageContainer extends StatelessWidget {
                 mainAxisAlignment:
                     isUser ? MainAxisAlignment.end : MainAxisAlignment.start,
                 mainAxisSize: MainAxisSize.min,
-                children: buttons,
+                children: buttons!,
               )
             else if (messageButtonsBuilder != null)
               Row(
                 crossAxisAlignment: CrossAxisAlignment.center,
                 mainAxisAlignment:
                     isUser ? MainAxisAlignment.end : MainAxisAlignment.start,
-                children: messageButtonsBuilder(message),
+                children: messageButtonsBuilder!(message),
                 mainAxisSize: MainAxisSize.min,
               ),
-            if (messageTimeBuilder != null)
-              messageTimeBuilder(
-                timeFormat != null
-                    ? timeFormat.format(message.createdAt)
-                    : DateFormat('HH:mm:ss').format(message.createdAt),
-                message,
-              )
-            else
-              Padding(
-                padding: EdgeInsets.only(top: 5.0),
-                child: Text(
-                  timeFormat != null
-                      ? timeFormat.format(message.createdAt)
-                      : DateFormat('HH:mm:ss').format(message.createdAt),
-                  style: TextStyle(
-                    fontSize: 10.0,
-                    color: message.user.color != null
-                        ? message.user.color
-                        : isUser
-                            ? Colors.white70
-                            : Colors.black87,
+            messageTimeBuilder?.call(
+                  timeFormat?.format(message.createdAt) ??
+                      DateFormat('HH:mm:ss').format(message.createdAt),
+                  message,
+                ) ??
+                Padding(
+                  padding: EdgeInsets.only(top: 5.0),
+                  child: Text(
+                    timeFormat != null
+                        ? timeFormat!.format(message.createdAt)
+                        : DateFormat('HH:mm:ss').format(message.createdAt),
+                    style: TextStyle(
+                      fontSize: 10.0,
+                      color: message.user.color ??
+                          (isUser ? Colors.white70 : Colors.black87),
+                    ),
                   ),
-                ),
-              )
+                )
           ],
         ),
       ),
@@ -175,38 +166,31 @@ class MessageContainer extends StatelessWidget {
   }
 
   Widget _buildMessageText() {
-    if (messageTextBuilder != null)
-      return messageTextBuilder(message.text, message);
-    else
-      return ParsedText(
-        parse: parsePatterns,
-        text: message.text,
-        style: TextStyle(
-          color: message.user.color != null
-              ? message.user.color
-              : isUser
-                  ? Colors.white70
-                  : Colors.black87,
-        ),
-      );
+    return messageTextBuilder?.call(message.text, message) ??
+        ParsedText(
+          parse: parsePatterns,
+          text: message.text!,
+          style: TextStyle(
+            color: message.user.color ??
+                (isUser ? Colors.white70 : Colors.black87),
+          ),
+        );
   }
 
   Widget _buildMessageImage() {
     if (message.image != null) {
-      if (messageImageBuilder != null)
-        return messageImageBuilder(message.image, message);
-      else
-        return Padding(
-          padding: EdgeInsets.only(top: 10.0, bottom: 10.0),
-          child: FadeInImage.memoryNetwork(
-            height: constraints.maxHeight * 0.3,
-            width: constraints.maxWidth * 0.7,
-            fit: BoxFit.contain,
-            placeholder: kTransparentImage,
-            image: message.image,
-          ),
-        );
+      return messageImageBuilder?.call(message.image, message) ??
+          Padding(
+            padding: EdgeInsets.only(top: 10.0, bottom: 10.0),
+            child: FadeInImage.memoryNetwork(
+              height: constraints!.maxHeight * 0.3,
+              width: constraints!.maxWidth * 0.7,
+              fit: BoxFit.contain,
+              placeholder: kTransparentImage,
+              image: message.image!,
+            ),
+          );
     }
-    return Container(width: 0, height: 0);
+    return SizedBox(width: 0, height: 0);
   }
 }
